docs: refactored the dropdownable component story issue number #1004
